+++ src/types.ts
@@ -1,0 +1,6 @@
+import type { DecorationOptions } from 'vscode'
+
+export interface DecorationMatch extends DecorationOptions {
+  content: string,
+  inline: boolean,
+}